SERVER-17282 update find command parsing to conform to find command spec changes
Also updates the shell to construct the find command according to the new spec.
This means that explain with special options (e.g. snapshot()) will not be compatible
when mixing 3.2 shell / 3.0 server or 3.0 shell / 3.2 server. Explain of find with
standard options like filter, query, projection, sort, and hint will be both backwards
and forwards compatible.

+++ src/mongo/s/commands/cluster_find_cmd.cpp
@@ -108,7 +108,7 @@ namespace mongo {
             vector<Strategy::CommandResult> shardResults;
             STRATEGY->commandOp(dbname,
                                 explainCmdBob.obj(),
-                                lpq->getOptions().toInt(),
+                                lpq->getOptions(),
                                 fullns,
                                 lpq->getFilter(),
                                 &shardResults);
